% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_on_fraction.R
\name{add_on_fraction}
\alias{add_on_fraction}
\title{Perfom annova on the data}
\usage{
add_on_fraction(data, threshold)
}
\arguments{
\item{data}{The data normalized}

\item{threshold}{The ON_OFF threshold}
}
\value{
An data table with on fractions
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
