% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_on_fraction.R
\name{compute_model_values}
\alias{compute_model_values}
\title{Compute BIC and loglike value of gaussian models}
\usage{
compute_model_values(data)
}
\arguments{
\item{data}{The data table}
}
\value{
An data table with BIC and loglik for each model
}
\description{
This function uses the Model-Based Clustering function Mclust to extract the optimal BIC value and the corresponding log-likelihood for a 1-gaussian model and a 2-gaussian model.
}
