% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dyn_profile.R
\name{density_value}
\alias{density_value}
\title{Estimates ordinate value of the density at position x.}
\usage{
density_value(x, d)
}
\arguments{
\item{x}{a numerical value}

\item{d}{a density function, as returned by density()}
}
\value{
the estimated ordinate value of the density at position x
}
\description{
Estimates ordinate value of the density at position x.
}
