% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dyn_profile.R
\name{draw.one.density}
\alias{draw.one.density}
\title{Translate density into colour, for graphix representation.}
\usage{
draw.one.density(d, ylim = range(d$x), xlim = c(0, 1), nstripes = 1000,
  colors = c("white", "darkblue"), ncolors = 100)
}
\arguments{
\item{d}{A density function, as returned by density()}

\item{ylim}{The limite of the drawned box on the Y axis, default is range(d$x)}

\item{xlim}{The limite of the drawned box on the X axis, default is c(0,1)}

\item{nstripes}{(integer) The number of horizontal stripes to plot per experiment, default is 1000}

\item{colors}{The colors to use to construct the ramp, default is c("white", "darkblue")}

\item{ncolors}{(integer) The number of gradual colors of the ramp, default is 100}
}
\value{
The colors and coordinates to use to plot a given density
}
\description{
Translate density into colour, for graphix representation.
}
\examples{
d = density(x=1:1000)
ylim = range(d$x)
xlim = c(0,1)
nstripes = 1000
colors = c("white", "darkblue"),
ncolors = 100
plot(1:2, xlim = xlim, ylim = ylim, type = "n")
draw.one.density (d, ylim, xlim, nstripes, colors, ncolors)
}
