% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dyn_profile.R
\name{dyn.profile}
\alias{dyn.profile}
\title{Draws a list of densities according to an index of color.}
\usage{
dyn.profile(ld, colors = c("white", "darkblue"), ncolors = 100,
  nstripes = 1000, xpos = NULL, ylim = NULL, ...)
}
\arguments{
\item{ld}{A list of density functions}

\item{colors}{The colors to use to construct the ramp, default is c("white", "darkblue")}

\item{ncolors}{(integer) The number of gradual colors of the ramp, default is 100}

\item{nstripes}{(integer) The number of horizontal stripes to plot per experiment, default is 1000}

\item{xpos}{A vector of positions on the x-axis to plot each experiment.}

\item{...}{Additional arguments to pass to the plot() function}
}
\value{
A plot of a serie of density profiles according to an index of color
}
\description{
Draws a list of densities according to an index of color.
}
\examples{
ld = list(density(x=1:1000), density(x=1:100), density(x=1:10))
ylim = range(d$x)
xlim = c(0,1)
nstripes = 1000
colors = c("white", "darkblue"),
ncolors = 100
dyn.profile (ld)
}
