% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_data}
\alias{export_data}
\title{Export final data}
\usage{
export_data(data)
}
\arguments{
\item{data}{The data to be exported}
}
\value{
A data tab
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
