% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facsticor.R
\docType{package}
\name{facsticor}
\alias{facsticor}
\alias{facsticor-package}
\title{facsticor: A package to analyse time course FACS data.}
\description{
The facsticor package provides three categories of important functions:
read_data, gate_data and dyn_profile.
}
\section{read_data functions}{

The read_data functions ...
}

\section{gate_data functions}{

The read_data functions ...
}

