% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_qc.R
\name{filter_data}
\alias{filter_data}
\title{Filters the data according to gating informations}
\usage{
filter_data(data, idx_to_be_removed)
}
\arguments{
\item{data}{The gated data generated by gate_facs_yeast_HTS_data}

\item{idx}{The index data to be removed}
}
\value{
A filtered data list
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
\examples{
load("data/data.rda")
idx_to_be_removed = (data$anno$nb_gcells < 500 | (data$anno$median_gssc < 33 & data$anno$plate=="PL3") )
data = filter_data(data, idx_to_be_removed)
}
