% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_on_fraction.R
\name{fit_ON_gaussian}
\alias{fit_ON_gaussian}
\title{Fit ON gaussian on the data}
\usage{
fit_ON_gaussian(on_cells)
}
\arguments{
\item{on_cells}{The ID of well in which cells are supposed to be on}
}
\value{
An data table with the number of cell on and the number of cell off
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
