% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_data.R
\name{gate_facs_yeast_HTS_data}
\alias{gate_facs_yeast_HTS_data}
\title{Interrogates the FSC/SSC density plot and extract ncell from the peak of density.}
\usage{
gate_facs_yeast_HTS_data(data, frac_cells = 0.4, graphics = TRUE,
  pre_gating_rm_rate = 0.05, verbose = TRUE, hard_lim = NULL)
}
\arguments{
\item{data}{A list containing a field called alldat, generated by the read_data.R function}

\item{pre_gating_rm_rate}{Pre-gating fraction of cells that will be remove on x and y axis. This allows to limit computations and filtering to the area of interest. Yeast cells being small, default value is 0.05.}

\item{frac_cell}{A fraction of cells to keep after gating (approximately). This number defines the size of the gate (few cells: narrow gate).}

\item{graphic}{If TRUE (default), graphics are produced in pdf files to visualize gating. The name of these files will be in the format platewell.pdf.}

\item{graphicsdir}{A character string giving the path to the directory where the graphics files should be written}
}
\value{
A list of data.frames containing the data of GATED cells. data$gdat in a list of gated data, and data$anno is the summary of the informations, it now contains the summary of gated data.
}
\description{
This function works of FL1H, FSC and SSC data acquired on 96-well plates
}
