% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_data.R
\name{gate_it}
\alias{gate_it}
\title{Generates the kerneld to gate the data}
\usage{
gate_it(x, y, idx, thres_1, thres_2, n = 34, DEBUG = FALSE)
}
\arguments{
\item{x}{A vector of FSC data}

\item{y}{A vector of SSC data}

\item{thres_1}{Pre-gating fraction of cells that will be remove on x and y axis. This allows to limit computations and filtering to the area of interest. Yeast cells being small, default value is 0.05.}

\item{thres_2}{A fraction of cells to keep after gating (approximately). This number defines the size of the gate (few cells: narrow gate).}
}
\value{
A kerneld
}
\description{
This function works of FL1H, FSC and SSC data acquired on 96-well plates
}
