% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_on_fraction.R
\name{get_probas}
\alias{get_probas}
\title{Perfom annova on the data}
\usage{
get_probas(data, mu_on, var_on, mu_off, var_off)
}
\arguments{
\item{data}{The data normalized}

\item{on_cells}{Cell used to modelized on cells}

\item{off_cells}{Cell used to modelized off cells}
}
\value{
An data table with the number of cell on and the number of cell off
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
