% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_on_fraction.R
\name{get_threshold}
\alias{get_threshold}
\title{Perfom annova on the data}
\usage{
get_threshold(t, eps, mu_on, sd_on, mu_off, sd_off)
}
\arguments{
\item{eps}{eps}

\item{mu_on}{The mean of ON cells}

\item{sd_on}{The standart deviation of ON cells}

\item{mu_off}{The standart deviation of OFF cells}

\item{data}{The data normalized}

\item{sd_on}{The variance of OFF cells}
}
\value{
An data table with the number of cell on and the number of cell off
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
