% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{perform_anova}
\alias{perform_anova}
\title{Perfom annova on the data}
\usage{
perform_anova(data, model)
}
\arguments{
\item{data}{A particular element of the list data$gdat (coresponds to wells present in all plates for normalization)}

\item{model}{The model to be used}
}
\value{
An annova model
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
\examples{
à finaliser
load("data/data.rda")
idx_to_be_keep = data$anno$rep\%in\%2:5 & data$anno$strain != "GY6" & TRUE
model_data = filter_data(data, !idx_to_be_keep)
model_p.s.t = mean_gfl1~plate+strain+time
}
