% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_qc.R
\name{plot_correl}
\alias{plot_correl}
\title{Plot the correlation between  FL1H, FSC and SSC}
\usage{
plot_correl(gated_data, main)
}
\arguments{
\item{gated_data}{A particular element of the list data$gdat (correspond to a single well)}
}
\value{
A figure of 3 plots to control gating efficiency
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
\examples{
load("data/data.rda")
gated_data=data$gdat[[1]]
cur_anno = data$anno[1, ]
main = paste(cur_anno$plate, cur_anno$well, sep = "")
plot_correl(gated_data, main)
}
