% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{plot_ctrl_norm}
\alias{plot_ctrl_norm}
\title{Plot the time course distributions accoording to mean FL1 after et before normalisation.}
\usage{
plot_ctrl_norm(data, idx_data, col)
}
\arguments{
\item{data}{The gated data generated by gate_facs_yeast_HTS_data}

\item{idx_data}{The index of the data that will be ploted}

\item{col}{The color index}
}
\value{
A plot of mean FL1 according to time
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
