% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_qc.R
\name{plot_gates}
\alias{plot_gates}
\title{Plot the gates}
\usage{
plot_gates(data, idx, xlim, ylim, col_key = "col", ...)
}
\arguments{
\item{data}{The gated data generated by gate_facs_yeast_HTS_data}

\item{idx}{The index of the data that will be ploted}

\item{col_key}{Which color indicator of the data$anno file will be used}
}
\value{
A figure of gates for indicated index
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
\examples{
load("data/data.rda")
col_key="col_time"
plot_gates(data, idx=data$anno$plate=="PL1", col_key=col_key)
}
