% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{plot_mean_fl1}
\alias{plot_mean_fl1}
\title{Plot the time course distributions accoording to mean FL1.}
\usage{
plot_mean_fl1(data, idx_data, type = "norm", col, lty = 1)
}
\arguments{
\item{data}{The gated data generated by gate_facs_yeast_HTS_data}

\item{idx_data}{The index of the data that will be ploted}

\item{type}{The type of data to be used, has to be either "norm" (by default) or "not_norm"}

\item{col}{The color index}

\item{lty}{The lty index, 1 by default}
}
\value{
A plot of mean FL1 according to time
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
