% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{plot_on_fraction}
\alias{plot_on_fraction}
\title{Plot the time course distributions accoording to the fraction of cell ON.}
\usage{
plot_on_fraction(data, idx_data, col, lty = 1)
}
\arguments{
\item{data}{The gated data generated by gate_facs_yeast_HTS_data}

\item{idx_data}{The index of the data that will be ploted}

\item{col}{The color index}

\item{lty}{The lty index, 1 by default}

\item{type}{The type of data to be used, has to be either "norm" (by default) or "not_norm"}
}
\value{
A plot of On frqction according to time
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
