% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_qc.R
\name{plot_qc}
\alias{plot_qc}
\title{Plot FL1H, FSC and SSC for each well}
\usage{
plot_qc(data, idx, type = "raw", col_key = "col", ...)
}
\arguments{
\item{data}{The gated data generated by gate_facs_yeast_HTS_data}

\item{idx}{The index of the data that will be ploted}

\item{type}{The type of data to be used, has to be either "raw" or "gated"}
}
\value{
A figure of 3 plots to control gating efficiency
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
\examples{
load("data/data.rda")
col_key="col_time"
plot_qc(data, idx=data$anno$plate=="PL1", type="raw", col_key=col_key)
}
