% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{plot_tc_mean}
\alias{plot_tc_mean}
\title{Plot the time course distributions accoording to color codes.}
\usage{
plot_tc_mean(data, idx_data, colors = c("white", "darkblue"), col = 1,
  lty = 1)
}
\arguments{
\item{data}{The gated data generated by gate_facs_yeast_HTS_data}

\item{idx_data}{The index of the data that will be ploted}

\item{col}{The color index}
}
\value{
A plot of distribution according to time
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates
}
