% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_facs_data}
\alias{read_facs_data}
\title{Reads the facs data get the summary informations on data distributions.}
\usage{
read_facs_data(datadir, datadate, annofile, anno = NULL)
}
\arguments{
\item{datadir}{The path to root directory of data}

\item{datadate}{The name of directory containing data, in date format: YYYMMDD, can be a vector of dates}

\item{annofile}{The experimenter annotation file, semi-colon separated, must at least contain the following informations :}
}
\value{
A list of data.frames containing the data of ALL cells. data$alldat in a list of data, and data$anno is the summary of the informations.
}
\description{
This function works on FL1H, FSC and SSC data acquired on 96-well plates. The data directory should be organized as follow: "date/plate/dat/your_fsc_data"
}
