% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dyn_profile.R
\name{x2i}
\alias{x2i}
\title{Indexes bins.}
\usage{
x2i(x, n)
}
\arguments{
\item{x}{A value between 0 and 1}

\item{n}{An integer indicating the number of bins between 0 and 1}
}
\value{
The index of the bin containing x, or NA if x > 1 or x > 0
}
\description{
Indexes bins.
}
